Split Pylint wrapper test support
Move fake Astroid modules, object-builder doubles, and routing
orchestration helpers into a dedicated test helper module so the
Pylint wrapper test file stays below the 400-line ceiling.
